pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eragent"
version = "0.1.0"
description = "ERP Analysis Agent based on LangChain 1.2.0 and Ontology"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "owlready2>=0.46",
    "neo4j>=5.0.0",
    "chromadb>=0.5.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "python-dateutil>=2.9.0",
    "faker>=25.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
include = ["api*", "core*", "config*", "modules*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=. --cov-report=term-missing --cov-fail-under=85"

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "*/__init__.py", "*/mock_data/*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
